+++ fabric/src/main/java/net/mehvahdjukaar/supplementaries/fabric/SupplementariesFabric.java
@@ -1,15 +1,14 @@
 package net.mehvahdjukaar.supplementaries.fabric;
 
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
-import net.fabricmc.fabric.api.client.screen.v1.ScreenKeyboardEvents;
 import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
-import net.mehvahdjukaar.moonlight.fabric.MLFabricSetupCallbacks;
 import net.mehvahdjukaar.supplementaries.Supplementaries;
 import net.mehvahdjukaar.supplementaries.common.events.fabric.ClientEventsFabric;
 import net.mehvahdjukaar.supplementaries.common.events.fabric.ServerEventsFabric;
 import net.mehvahdjukaar.supplementaries.common.utils.VibeChecker;
 import net.mehvahdjukaar.supplementaries.reg.ModSetup;
+import net.minecraft.core.RegistryCodecs;
+import net.minecraft.core.registries.Registries;
 
 public class SupplementariesFabric implements ModInitializer {
 
@@ -22,11 +21,12 @@ public class SupplementariesFabric implements ModInitializer {
         if (PlatHelper.getPhysicalSide().isClient()) {
             VibeChecker.checkVibe();
             ClientEventsFabric.init();
-            SupplementariesFabricClient.clientInitAndSetup();
+
+            SupplementariesFabricClient.init();
         }
 
-        MLFabricSetupCallbacks.COMMON_SETUP.add(ModSetup::setup);
-        MLFabricSetupCallbacks.COMMON_SETUP.add(ModSetup::asyncSetup);
+        PlatHelper.addCommonSetup(ModSetup::setup);
+        PlatHelper.addCommonSetup(ModSetup::asyncSetup);
 
     }
 }